get accent color from recoil state
