+++ src/components/expanded-state/ContactProfileState.js
@@ -1,6 +1,7 @@
 import lang from 'i18n-js';
 import React, { useCallback, useRef, useState } from 'react';
 import { Keyboard } from 'react-native';
+import { useRecoilValue } from 'recoil';
 import { useTheme } from '../../context/ThemeContext';
 import { useNavigation } from '../../navigation/Navigation';
 import { abbreviations, magicMemo, profileUtils } from '../../utils';
@@ -20,6 +21,7 @@ import {
   removeFirstEmojiFromString,
   returnStringFirstEmoji,
 } from '@rainbow-me/helpers/emojiHandler';
+import { accentColorAtom } from '@rainbow-me/helpers/ens';
 import { isValidDomainFormat } from '@rainbow-me/helpers/validators';
 import { useAccountSettings, useContacts } from '@rainbow-me/hooks';
 import { ImgixImage } from '@rainbow-me/images';
@@ -87,7 +89,6 @@ const centerdStyles = padding.object(24, 25);
 const bottomStyles = padding.object(8, 9);
 
 const ContactProfileState = ({
-  accentColor,
   address,
   avatarUrl,
   contact,
@@ -98,6 +99,10 @@ const ContactProfileState = ({
   const contactNickname = contact?.nickname || nickname;
   const { goBack } = useNavigation();
   const { onAddOrUpdateContacts, onRemoveContact } = useContacts();
+  const { isDarkMode, colors } = useTheme();
+
+  const accentColor = useRecoilValue(accentColorAtom);
+
   const [color, setColor] = useState(accentColor || 0);
   const [value, setValue] = useState(
     removeFirstEmojiFromString(contactNickname)
@@ -139,8 +144,6 @@ const ContactProfileState = ({
   ]);
 
   const isContact = Boolean(contact);
-
-  const { isDarkMode, colors } = useTheme();
 
   const handleChangeAvatar = useCallback(() => {
     const prevAvatarIndex = profileUtils.avatars.findIndex(
